added dropdown in add expense & limited display
the "add expense" modal now presents a dynamically created dropdown menu of the categories, rather than having the user type them in;

the categories list/display now only shows 3 categories by default. If more categories are added, a "show more" button appears to expand the list. when the list is expanded, a "show less" button appears to collapse it again if clicked.